refactor: Move share via QR into its own component
